Load member config from YAML
Load member config from YAML similarly to XML either from:
- Classpath
- File system
- In-memory stream

The same config file location mechanism is in use as for XML (YamlConfigLocator).
If the HZ instance is built with no config, an attempt is made to read the config from any provided YAML configuration (with excluding default
YAML configuration). If there is no YAML configuration provided, the configuration is loaded from XML as before, this time including the default
XML as the last option.

Added yaml configs to the classpath:
- hazelcast-default.yaml
- hazelcast-fullconfig.yaml (test)
- hazelcast-service.yaml (test)

Along with the changes missing configuration is added to hazelcast-fullconfig.xml as well.

Notable tests:
- YamlConfigBuilderTest with cases identical to the ones in XMLConfigBuilderTest (schema validation cases are ignored for now)
- Equality test for the default YAML and XML configs provided on the classpath
- Equality test for the full YAML and XML configs provided on the (test) classpath

While this is a change focusing on member configuration, contains some changes in client config classes as well where it is
needed to compile.