Replace and import support in YAML member config
Add support for importing external documents and replacing variables in
YAML member configs.

Replacing is fully identical to the behavior of replacing in XML.

Importing has one difference. Importing XML documents replaces the
import tag with the content of the imported XML document. In the case of
YAML replacing is not enough, the two YAML documents need to be merged
recursively in order to ensure we get a valid YAML document as the
result. This is done at YAML DOM level, following this logic:
- if a given source node is not found in the target DOM, we attach the
given node to the target
- if the node is found, we recursively continue the process with the
given node

This results in a difference if the same datastructure config is spanned
over two files. In the case of YAML, the result configuration is merged,
taking the configuration values from both files. In the XML case, the
configuration values taken from the main XML file - the one that imports
the other XML. This can be seen in the two implementations of the
AbstractConfigImportVariableReplacementTest#testMapConfigFromMainAndImportedFile.
The YAML merging logic is implemented in YamlConfigBuilder#merge.